Increase Netex version size

+++ mobi.chouette.exchange.neptune/src/main/java/mobi/chouette/exchange/neptune/parser/ConnectionLinkParser.java
@@ -49,7 +49,7 @@ public class ConnectionLinkParser implements Parser, Constant {
 						objectId);
 				connectionLink.setFilled(true);
 			} else if (xpp.getName().equals("objectVersion")) {
-				Integer version = ParserUtils.getInt(xpp.nextText());
+				Long version = ParserUtils.getLong(xpp.nextText());
 				connectionLink.setObjectVersion(version);
 			} else if (xpp.getName().equals("creationTime")) {
 				LocalDateTime creationTime = ParserUtils.getLocalDateTime(xpp.nextText());
